Add files via upload
A CRUD application is a type of software application that performs basic operations on data. CRUD stands for:
C – Create
R - Read
U -Update
D -Delete
which represent the four basic functions that can be performed on persistent storage. These operations allow users to interact with and manage data within a system.